requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
python-docx
openpyxl
